backend/requirements.txt
flask
python-dotenv
google-generativeai>=0.3.0
flask-cors